moved files around and made the serializable class abstract
