+++ pgsqltoolsservice/admin/contracts/get_databaseinfo_request.py
@@ -4,7 +4,7 @@
 # --------------------------------------------------------------------------------------------
 
 from pgsqltoolsservice.hosting import IncomingMessageConfiguration
-from pgsqltoolsservice.common import Serializable
+from pgsqltoolsservice.serialization import Serializable
 
 
 class GetDatabaseInfoParameters(Serializable):
