+++ pgsqltoolsservice/capabilities/contracts/initialize_request.py
@@ -6,7 +6,7 @@
 import enum
 
 from pgsqltoolsservice.hosting import IncomingMessageConfiguration
-from pgsqltoolsservice.common import Serializable
+from pgsqltoolsservice.serialization import Serializable
 
 
 class InitializeRequestParams(Serializable):
