+++ pgsqltoolsservice/connection/contracts/cancellation_request.py
@@ -5,7 +5,7 @@
 
 from pgsqltoolsservice.connection.contracts.common import ConnectionType
 from pgsqltoolsservice.hosting import IncomingMessageConfiguration
-from pgsqltoolsservice.common import Serializable
+from pgsqltoolsservice.serialization import Serializable
 
 
 class CancelConnectParams(Serializable):
