+++ pgsqltoolsservice/connection/contracts/connect_request.py
@@ -5,7 +5,7 @@
 
 from pgsqltoolsservice.hosting import IncomingMessageConfiguration
 from pgsqltoolsservice.connection.contracts.common import ConnectionDetails, ConnectionType  # noqa
-from pgsqltoolsservice.common import Serializable
+from pgsqltoolsservice.serialization import Serializable
 
 
 class ConnectRequestParams(Serializable):
