+++ pgsqltoolsservice/connection/contracts/list_databases.py
@@ -8,7 +8,7 @@
 from typing import List  # noqa
 
 from pgsqltoolsservice.hosting import IncomingMessageConfiguration
-from pgsqltoolsservice.common import Serializable
+from pgsqltoolsservice.serialization import Serializable
 
 
 class ListDatabasesParams(Serializable):
