+++ pgsqltoolsservice/disaster_recovery/contracts/backup.py
@@ -9,7 +9,7 @@ import enum
 from typing import List  # noqa
 
 from pgsqltoolsservice.hosting import IncomingMessageConfiguration
-from pgsqltoolsservice.common import Serializable
+from pgsqltoolsservice.serialization import Serializable
 
 
 class BackupParams(Serializable):
