+++ pgsqltoolsservice/language/contracts/intellisense_ready.py
@@ -6,7 +6,7 @@
 """This module holds contracts for the IntelliSense Ready notification"""
 
 from pgsqltoolsservice.hosting import IncomingMessageConfiguration
-from pgsqltoolsservice.common import Serializable
+from pgsqltoolsservice.serialization import Serializable
 
 
 class IntelliSenseReadyParams(Serializable):
