+++ pgsqltoolsservice/query_execution/contracts/execute_request.py
@@ -5,7 +5,7 @@
 
 from pgsqltoolsservice.hosting import IncomingMessageConfiguration
 from pgsqltoolsservice.query_execution.contracts.common import SelectionData
-from pgsqltoolsservice.common import Serializable
+from pgsqltoolsservice.serialization import Serializable
 
 
 class ExecuteRequestParamsBase(Serializable):
