+++ pgsqltoolsservice/query_execution/contracts/query_request.py
@@ -5,7 +5,7 @@
 
 
 from pgsqltoolsservice.hosting import IncomingMessageConfiguration
-from pgsqltoolsservice.common import Serializable
+from pgsqltoolsservice.serialization import Serializable
 
 
 class SubsetParams(Serializable):
