+++ pgsqltoolsservice/workspace/contracts/did_open_text_doc_notification.py
@@ -5,7 +5,7 @@
 
 from pgsqltoolsservice.hosting import IncomingMessageConfiguration
 from pgsqltoolsservice.workspace.contracts.common import TextDocumentItem
-from pgsqltoolsservice.common import Serializable
+from pgsqltoolsservice.serialization import Serializable
 
 
 class DidOpenTextDocumentParams(Serializable):
